pyproject.toml
[tool.poetry]
name = "phala-cloud-deployment-action"
version = "0.1.0"
description = "A script to deploy CVMs to Phala Cloud, intended for use in a GitHub Action."
authors = ["Volodymyr Isai <volod@opendatalabs.xyz>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
cryptography = "^42.0.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"